+++ llmfoundry/data_prep/__init__.py
@@ -1,0 +1,12 @@
+# Copyright 2024 MosaicML LLM Foundry authors
+# SPDX-License-Identifier: Apache-2.0
+
+from llmfoundry.data_prep.convert_dataset_json import (
+    convert_dataset_json,
+    convert_dataset_json_from_args,
+)
+
+__all__ = [
+    'convert_dataset_json',
+    'convert_dataset_json_from_args',
+]
