+++ llmfoundry/data_prep/convert_dataset_json.py
@@ -1,0 +1,163 @@
+# Copyright 2022 MosaicML LLM Foundry authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Streaming dataset conversion scripts for json files."""
+import os
+from argparse import Namespace
+from enum import Enum
+from glob import glob
+from typing import Optional
+
+import datasets as hf_datasets
+from streaming import MDSWriter
+from torch.utils.data import IterableDataset
+from tqdm import tqdm
+from transformers import AutoTokenizer, PreTrainedTokenizerBase
+
+from llmfoundry.data import ConcatTokensDataset, NoConcatDataset
+
+
+class ConcatMode(Enum):
+    NO_CONCAT = 'NO_CONCAT'
+    CONCAT_TOKENS = 'CONCAT_TOKENS'
+
+
+def build_hf_dataset(
+    path: str,
+    split: str,
+    mode: ConcatMode,
+    max_length: Optional[int] = None,
+    bos_text: str = '',
+    eos_text: str = '',
+    no_wrap: bool = False,
+    tokenizer: PreTrainedTokenizerBase = None,
+) -> IterableDataset:
+    """Build an IterableDataset over the HF C4 or pile source data.
+
+    Args:
+        dataset_name (str): Dataset name
+        split (str): Split name.
+        mode (ConcatMode): NO_CONCAT, or CONCAT_TOKENS
+        max_length (int): The length of concatenated tokens
+        bos_text (str): text to insert at the beginning of each sequence
+        eos_text (str): text to insert at the end of each sequence
+        no_wrap (bool): if concatenating, whether to wrap text across `max_length` boundaries
+        tokenizer (PreTrainedTokenizerBase): if mode is CONCAT_TOKENS, the tokenizer to use
+        data_subset (str): Referred to as "name" in HuggingFace datasets.load_dataset.
+            Typically "all" (The Pile) or "en" (c4).
+
+    Returns:
+        An IterableDataset.
+    """
+    if os.path.isdir(path):
+        data_files = glob(f'{path}/*')
+    else:
+        data_files = path
+
+    hf_dataset = hf_datasets.load_dataset(
+        'json',
+        data_files=data_files,
+        split=split,
+    )
+
+    if mode == ConcatMode.NO_CONCAT:
+        dataset = NoConcatDataset(hf_dataset)
+    else:
+        if not isinstance(tokenizer, PreTrainedTokenizerBase):
+            raise ValueError(
+                f'{tokenizer=} must be of type PreTrainedTokenizerBase',
+            )
+        if max_length is None:
+            raise ValueError(f'max_length must be set.')
+        if bos_text + eos_text == '':
+            test_tokens = tokenizer('test')
+            if test_tokens['input_ids'][
+                0] != tokenizer.bos_token_id and test_tokens['input_ids'][
+                    -1] != tokenizer.eos_token_id:
+                tok_error_msg = 'This tokenizer does not insert an EOS nor BOS token. '
+                tok_error_msg += 'Concatenating with this tokenizer will result in sequences being '
+                tok_error_msg += 'attached without a separating token. Please use another tokenizer, '
+                tok_error_msg += 'such as facebook/opt-125m, or specify EOS/BOS text with e.g. '
+                tok_error_msg += '--bos_text=<|endoftext|>.'
+                raise ValueError(tok_error_msg)
+        dataset = ConcatTokensDataset(
+            hf_dataset=hf_dataset,
+            tokenizer=tokenizer,
+            max_length=max_length,
+            bos_text=bos_text,
+            eos_text=eos_text,
+            no_wrap=no_wrap,
+        )
+    return dataset
+
+
+def convert_dataset_json(args: Namespace) -> None:
+    """Main: create C4/pile streaming dataset.
+
+    Args:
+        args (Namespace): Commandline arguments.
+    """
+    if args.concat_tokens is not None:
+        mode = ConcatMode.CONCAT_TOKENS
+        tokenizer = AutoTokenizer.from_pretrained(args.tokenizer)
+        # we will enforce length, so suppress warnings about sequences too long for the model
+        tokenizer.model_max_length = int(1e30)
+        columns = {'tokens': 'ndarray:int32'}
+    else:
+        mode = ConcatMode.NO_CONCAT
+        tokenizer = None
+        columns = {'text': 'str'}
+
+    # Get samples
+    dataset = build_hf_dataset(
+        path=args.path,
+        split=args.split,
+        mode=mode,
+        max_length=args.concat_tokens,
+        bos_text=args.bos_text,
+        eos_text=args.eos_text,
+        no_wrap=args.no_wrap,
+        tokenizer=tokenizer,
+    )
+
+    print('here')
+
+    # Write samples
+    print(f'Converting to MDS format...')
+    print(
+        f'Note that the progress bar is based on the dataset length before tokenization.',
+    )
+    print(f'It will finish at a value below 100% if tokenizing')
+    with MDSWriter(
+        columns=columns,
+        out=os.path.join(args.out_root),
+        compression=args.compression,
+    ) as out:
+        for sample in tqdm(dataset):
+            out.write(sample)
+
+
+def convert_dataset_json_from_args(args: Namespace) -> None:
+    if os.path.isdir(args.out_root) and len(
+        set(os.listdir(args.out_root)).intersection(set(args.split)),
+    ) > 0:
+        raise ValueError(
+            f'--out_root={args.out_root} contains {os.listdir(args.out_root)} which cannot overlap with the requested splits {args.splits}.',
+        )
+
+    # Make sure we have needed concat options
+    if (
+        args.concat_tokens is not None and
+        isinstance(args.concat_tokens, int) and args.tokenizer is None
+    ):
+        args.error(
+            'When setting --concat_tokens, you must specify a --tokenizer',
+        )
+
+    # now that we have validated them, change BOS/EOS to strings
+    if args.bos_text is None:
+        args.bos_text = ''
+    if args.eos_text is None:
+        args.eos_text = ''
+
+    convert_dataset_json(args)
